+++ packages/component/src/BasicFilm.js
@@ -17,7 +17,7 @@ const CAROUSEL_CSS = css({
   position: 'relative'
 });
 
-export default class BasicFilm extends React.Component {
+class BasicFilm extends React.Component {
   constructor(props) {
     super(props);
 
@@ -43,55 +43,68 @@ export default class BasicFilm extends React.Component {
 
     const {
       leftFlipperText = '<',
+      numItems,
       rightFlipperText = '>',
+      scrollBarWidth,
+      scrolling,
       showDots = true,
       showFlipper = true,
       showScrollBar = true
     } = props;
 
     return (
-      <Composer>
-        <Context.Consumer>
-          { ({ numItems, scrollBarWidth, scrolling }) =>
-            <div className={ props.className }>
-              <div
-                className={ classNames(CAROUSEL_CSS + '', { scrolling }, carousel + '') }
-                style={ this.createHeightStyle(props.height) }
+      <React.Fragment>
+        <div className={ props.className }>
+          <div
+            className={ classNames(CAROUSEL_CSS + '', { scrolling }, carousel + '') }
+            style={ this.createHeightStyle(props.height) }
+          >
+            { !!numItems && scrollBarWidth !== '100%' && !!showFlipper &&
+              <Flipper className={ leftFlipper + '' } mode="left">
+                <div>{ leftFlipperText }</div>
+              </Flipper>
+            }
+            <FilmStrip>
+              { props.children }
+            </FilmStrip>
+            { !!numItems && scrollBarWidth !== '100%' && !!showFlipper &&
+              <Flipper className={ rightFlipper + '' } mode="right">
+                <div>{ rightFlipperText }</div>
+              </Flipper>
+            }
+            { !!numItems && scrollBarWidth !== '100%' && !!showScrollBar &&
+              <ScrollBar
+                className={ scrollBarBox + '' }
+                handlerClassName={ scrollBarHandler + '' }
+              />
+            }
+          </div>
+          {
+            !!numItems && scrollBarWidth !== '100%' && !!showDots &&
+              <Dots
+                className={ dotsBox + '' }
+                itemClassName={ dotsItem + ''}
               >
-                { !!numItems && scrollBarWidth !== '100%' && !!showFlipper &&
-                  <Flipper className={ leftFlipper + '' } mode="left">
-                    <div>{ leftFlipperText }</div>
-                  </Flipper>
-                }
-                <FilmStrip>
-                  { props.children }
-                </FilmStrip>
-                { !!numItems && scrollBarWidth !== '100%' && !!showFlipper &&
-                  <Flipper className={ rightFlipper + '' } mode="right">
-                    <div>{ rightFlipperText }</div>
-                  </Flipper>
-                }
-                { !!numItems && scrollBarWidth !== '100%' && !!showScrollBar &&
-                  <ScrollBar
-                    className={ scrollBarBox + '' }
-                    handlerClassName={ scrollBarHandler + '' }
-                  />
-                }
-              </div>
-              {
-                !!numItems && scrollBarWidth !== '100%' && !!showDots &&
-                  <Dots
-                    className={ dotsBox + '' }
-                    itemClassName={ dotsItem + ''}
-                  >
-                    { () => <div /> }
-                  </Dots>
-              }
-            </div>
+                { () => <div /> }
+              </Dots>
           }
-        </Context.Consumer>
+        </div>
         { props.autoCenter !== false && <AutoCenter /> }
-      </Composer>
+      </React.Fragment>
     );
   }
 }
+
+export default props =>
+  <Composer numItems={ React.Children.count(props.children) }>
+    <Context.Consumer>
+      { ({ numItems, scrollBarWidth, scrolling }) =>
+        <BasicFilm
+          { ...props }
+          numItems={ numItems }
+          scrollBarWidth={ scrollBarWidth }
+          scrolling={ scrolling }
+        />
+      }
+    </Context.Consumer>
+  </Composer>