+++ packages/component/src/Composer.js
@@ -1,17 +1,21 @@
+import memoize from 'memoize-one';
 import React from 'react';
 
 import best from './best';
 import Context from './Context';
-import InternalContext from './InternalContext';
 import ScrollSpy from './ScrollSpy';
 import ScrollTo from './ScrollTo';
 
-function getView(current, scrollingTo) {
-  if (current) {
-    const scrollLeft = scrollingTo || current.scrollLeft;
-    const items = current.children[0].children; // This will enumerate <li> inside <FilmStrip>
-    const scrollCenter = scrollLeft + current.offsetWidth / 2;
-    const index = best([].slice.call(items), item => {
+function getView(
+  { current: scrollable } = {},
+  { current: itemContainer } = {},
+  scrollingTo
+) {
+  if (itemContainer && scrollable) {
+    const scrollLeft = scrollingTo || scrollable.scrollLeft;
+    const items = itemContainer.children; // This will enumerate <li> inside <FilmStrip>
+    const scrollCenter = scrollLeft + scrollable.offsetWidth / 2;
+    const index = best([...items], item => {
       const offsetCenter = item.offsetLeft + item.offsetWidth / 2;
 
       return 1 / Math.abs(scrollCenter - offsetCenter);
@@ -23,10 +27,10 @@ function getView(current, scrollingTo) {
       let indexFraction = index + (scrollCenter - offsetCenter) / item.offsetWidth;
 
       // We "fix" indexFraction if the viewport is at the start/end of the content
-      // This is to simplify code that use Math.round(indexFraction) to find the current index
+      // This is to simplify code that use Math.round(indexFraction) to find the scrollable index
       // if (scrollLeft === 0) {
       //   indexFraction = 0;
-      // } else if (scrollLeft >= current.scrollWidth - current.offsetWidth) {
+      // } else if (scrollLeft >= scrollable.scrollWidth - scrollable.offsetWidth) {
       //   indexFraction = items.length - 1;
       // } else if (indexFraction % 1 > .99 || indexFraction % 1 < .01) {
       //   indexFraction = Math.round(indexFraction);
@@ -40,7 +44,7 @@ function getView(current, scrollingTo) {
 
       if (scrollLeft === 0) {
         selectedIndex = 0;
-      } else if (scrollLeft >= current.scrollWidth - current.offsetWidth) {
+      } else if (scrollLeft >= scrollable.scrollWidth - scrollable.offsetWidth) {
         selectedIndex = items.length - 1;
       } else {
         selectedIndex = Math.round(indexFraction);
@@ -48,23 +52,25 @@ function getView(current, scrollingTo) {
 
       return {
         index: selectedIndex,
-        indexFraction,
-        items,
-        current
+        indexFraction
       };
     }
   }
 }
 
-function getScrollLeft(current, index) {
-  if (current) {
-    const items = current.children[0].children; // This will enumerate <li> inside <FilmStrip>
+function getScrollLeft(
+  { current: scrollable } = {},
+  { current: itemContainer } = {},
+  index
+) {
+  if (itemContainer && scrollable) {
+    const items = itemContainer.children; // This will enumerate <li> inside <FilmStrip>
     const item = items[Math.max(0, Math.min(items.length - 1, index))];
 
     if (item) {
       const itemOffsetCenter = item.offsetLeft + item.offsetWidth / 2;
 
-      return itemOffsetCenter - current.offsetWidth / 2;
+      return itemOffsetCenter - scrollable.offsetWidth / 2;
     }
   }
 }
@@ -76,24 +82,31 @@ export default class FilmComposer extends React.Component {
     this.handleScroll = this.handleScroll.bind(this);
     this.handleScrollToEnd = this.handleScrollToEnd.bind(this);
 
+    this.itemContainerRef = React.createRef();
+    this.scrollableRef = React.createRef();
+
+    this.mergeContext = memoize((state, numItems = 0) => ({
+      ...state,
+      numItems
+    }));
+
     this.state = {
-      filmStrip: null,
-      scrollLeft: null,
       context: {
-        numItems: 0,
+        itemContainerRef: this.itemContainerRef,
+        scrollableRef: this.scrollableRef,
         scrollBarPercentage: '0%',
         scrollBarWidth: '0%',
         scrolling: false,
         scrollTo: scrollTo => {
           this.setState(state => {
-            const view = getView(state.filmStrip, state.scrollLeft);
+            const view = getView(this.scrollableRef, this.itemContainerRef, state.scrollLeft);
 
             if (view) {
               const { index, indexFraction } = view;
               const targetIndex = scrollTo({ index, indexFraction });
 
               if (typeof targetIndex === 'number') {
-                return { scrollLeft: getScrollLeft(state.filmStrip, targetIndex) };
+                return { scrollLeft: getScrollLeft(this.scrollableRef, this.itemContainerRef, targetIndex) };
               }
             }
           });
@@ -105,17 +118,7 @@ export default class FilmComposer extends React.Component {
           this.state.context.scrollTo(({ indexFraction }) => Math.floor(indexFraction) + 1);
         }
       },
-      internalContext: {
-        _setFilmStripRef: filmStrip => this.setState(() => ({ filmStrip })),
-        _setNumItems: numItems => {
-          this.setState(({ context }) => ({
-            context: {
-              ...context,
-              numItems
-            }
-          }));
-        },
-      }
+      scrollLeft: null
     };
   }
 
@@ -123,9 +126,13 @@ export default class FilmComposer extends React.Component {
     clearTimeout(this.scrollTimeout);
   }
 
-  handleScroll({ fraction: scrollBarPercentage, initial, width: scrollBarWidth }) {
-    this.setState(({ context, filmStrip, scrollLeft }) => {
-      const view = getView(filmStrip, scrollLeft);
+  handleScroll({
+    fraction: scrollBarPercentage,
+    initial,
+    width: scrollBarWidth
+  }) {
+    this.setState(({ context, scrollLeft }) => {
+      const view = getView(this.scrollableRef, this.itemContainerRef, scrollLeft);
 
       if (view) {
         const { index, indexFraction } = view;
@@ -145,6 +152,7 @@ export default class FilmComposer extends React.Component {
 
     if (!initial) {
       clearTimeout(this.scrollTimeout);
+
       this.scrollTimeout = setTimeout(() => {
         this.setState(({ context }) => ({
           context: {
@@ -161,31 +169,35 @@ export default class FilmComposer extends React.Component {
   }
 
   render() {
-    const { state } = this;
+    const {
+      props: {
+        children,
+        numItems
+      },
+      scrollableRef,
+      state: {
+        context,
+        scrollLeft
+      }
+    } = this;
 
     return (
-      <InternalContext.Provider value={ state.internalContext }>
-        <Context.Provider value={ state.context }>
-          { this.props.children }
-          {
-            !!state.filmStrip &&
-              <ScrollSpy
-                onScroll={ this.handleScroll }
-                target={ state.filmStrip }
-              />
-          }
-          {
-            typeof state.scrollLeft === 'number'
-            && !!state.filmStrip
-            &&
-              <ScrollTo
-                onEnd={ this.handleScrollToEnd }
-                scrollLeft={ state.scrollLeft }
-                target={ state.filmStrip }
-              />
-          }
-        </Context.Provider>
-      </InternalContext.Provider>
+      <Context.Provider value={ this.mergeContext(context, numItems) }>
+        { children }
+        <ScrollSpy
+          onScroll={ this.handleScroll }
+          targetRef={ scrollableRef }
+        />
+        {
+          typeof scrollLeft === 'number'
+          &&
+            <ScrollTo
+              onEnd={ this.handleScrollToEnd }
+              scrollLeft={ scrollLeft }
+              targetRef={ scrollableRef }
+            />
+        }
+      </Context.Provider>
     );
   }
 }